awaiting task fsm state
added the awaiting task state for the fsm, this allows the agent to request a task. Also removed the circular reference within the fsm and it's states by changing the way they interact. Added a function overload to the task builder class so that it can take a vector of strings which it can then use to get the agents that correspond to those keys, it then call the original function passing in that new vector.